backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.9.2
websockets==14.2
google-genai>=1.45.0

